misc * remove file only when file exists * add apache2-utils to use "htpasswd"

+++ recipes/nginx.rb
@@ -1,4 +1,5 @@
 package "nginx"
+package "apache2-utils"
 
 execute "backup old file" do
   command "cp -p /etc/nginx/nginx.conf /etc/nginx/nginx.conf.old"
@@ -6,6 +7,7 @@ end
 
 execute "remove default file" do
   command "rm /etc/nginx/sites-enabled/default"
+  only_if "test -e /etc/nginx/sites-enabled/default"
 end
 
 template "/etc/nginx/nginx.conf" do
